consumer/requirements.txt
kafka-python
hdfs
pyarrow